Modulated books.js into several individual files
