+++ index.html
@@ -19,7 +19,7 @@
       <button id="addBtn" type="button">Add</button>
    </main>
 
-   <script src="js/books.js"></script>
+   <script type="module" src="js/main.js"></script>
 </body>
 
 </html>
