+++ js/book.js
@@ -1,0 +1,7 @@
+export default class Book {
+  constructor(id, title, author) {
+    this.id = id;
+    this.title = title;
+    this.author = author;
+  }
+}
