+++ js/main.js
@@ -1,0 +1,4 @@
+import App from './app.js';
+
+const app = new App();
+app.start();
